feat: With these changes, the structure for the status of newly created blog posts and existing blog posts is updated.
With these changes, the structure for the status of newly created blog posts and existing blog posts is updated. For blog posts that have already been created with different statuses, we can now change the status of the post depending on the user role. A content manager can change the status from 'Draft' to 'In Review', while an admin can change the status to 'Published'. For newly created blog posts, if the user role is 'Admin', it shows the option to publish the blog post. Optionally, for better functionality of the application, when a blog post is successfully created, it redirects to the blogs route.