fix: Change in creating utils for Blog status, using assertion and avoiding of duplicate useState
With this change, we now have utilities for showing blog status with a capital letter and without using "_", making it easily reusable in other parts of the project. Additionally, there is a change in displaying status options; we are now using assertions, and statusOptions is immutable by declaring it as a const. Finally, there is an update to avoid using two useState in BlogDetailsCard. Instead, we are using useRef, where initialStatus will hold the status before any changes when editing the blog post.